backend/requirements.txt
flask
flask-cors
pyjwt
python-dotenv
gunicorn
